+++ src/Braintree/ApplePayDetails.cs
@@ -13,6 +13,17 @@ namespace Braintree
         public virtual string Token { get; protected set; }
         public virtual string LastFour { get; protected set; }
         public virtual string ImageUrl { get; protected set; }
+        public virtual string Prepaid { get; protected set; }
+        public virtual string Healthcare { get; protected set; }
+        public virtual string Debit { get; protected set; }
+        public virtual string DurbinRegulated { get; protected set; }
+        public virtual string Commercial { get; protected set; }
+        public virtual string Payroll { get; protected set; }
+        public virtual string IssuingBank { get; protected set; }
+        public virtual string CountryOfIssuance { get; protected set; }
+        public virtual string ProductId { get; protected set; }
+        public virtual string Bin { get; protected set; }
+        public virtual string GlobalId { get; protected set; }
 
         protected internal ApplePayDetails(NodeWrapper node)
         {
@@ -25,6 +36,17 @@ namespace Braintree
             Token = node.GetString("token");
             LastFour = node.GetString("last-4");
             ImageUrl = node.GetString("image-url");
+            Prepaid = node.GetString("prepaid");
+            Healthcare = node.GetString("healthcare");
+            Debit = node.GetString("debit");
+            DurbinRegulated = node.GetString("durbin-regulated");
+            Commercial = node.GetString("commercial");
+            Payroll = node.GetString("payroll");
+            IssuingBank = node.GetString("issuing-bank");
+            CountryOfIssuance = node.GetString("country-of-issuance");
+            ProductId = node.GetString("product-id");
+            Bin = node.GetString("bin");
+            GlobalId = node.GetString("global-id");
         }
 
         [Obsolete("Mock Use Only")]